+++ app/src/main/java/revolut/converter/data/datasource/local/preferences/PreferencesHelper.kt
@@ -2,5 +2,5 @@ package revolut.converter.data.datasource.local.preferences
 
 interface PreferencesHelper {
     var currencyCode: String
-    var amount: Float
+    var amount: Double
 }
