+++ app/src/main/java/revolut/converter/data/datasource/remote/model/RatesList.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class RatesList(
         @SerializedName("baseCurrency") val baseCurrency: String,
-        @SerializedName("rates") val currencies: Map<String, Float>
+        @SerializedName("rates") val currencies: Map<String, Double>
 )
